+++ ai-services/cmd/ai-services/cmd/mustgather/mustgather.go
@@ -15,7 +15,20 @@ var (
 	runtimeType     string
 	outputDir       string
 	applicationName string
+	namespace       string
 )
+
+// gatherOptions carries options forwarded from the cobra command.
+type gatherOptions struct {
+	outputDir       string
+	applicationName string
+	namespace       string
+}
+
+// gatherer is the common interface implemented by every runtime-specific collector.
+type gatherer interface {
+	gather(opts gatherOptions) (string, error)
+}
 
 // MustGatherCmd returns the must-gather cobra command.
 func MustGatherCmd() *cobra.Command {
@@ -73,18 +86,18 @@ func mustGatherPreRun(cmd *cobra.Command, _ []string) error {
 func mustGatherRun(cmd *cobra.Command, _ []string) error {
 	cmd.SilenceUsage = true
 
-	rt := vars.RuntimeFactory.GetRuntimeType()
-	if rt != types.RuntimeTypePodman {
-		return fmt.Errorf("must-gather currently supports only the 'podman' runtime")
-	}
-
-	gatherer := newPodmanGatherer()
 	opts := gatherOptions{
 		outputDir:       outputDir,
 		applicationName: applicationName,
+		namespace:       namespace,
 	}
 
-	outDir, err := gatherer.gather(opts)
+	g, err := newGatherer(vars.RuntimeFactory.GetRuntimeType())
+	if err != nil {
+		return err
+	}
+
+	outDir, err := g.gather(opts)
 	if err != nil {
 		return fmt.Errorf("must-gather failed: %w", err)
 	}
@@ -92,4 +105,14 @@ func mustGatherRun(cmd *cobra.Command, _ []string) error {
 	logger.Infof("Must-gather complete. Output saved to: %s\n", outDir)
 
 	return nil
+}
+
+// newGatherer returns the gatherer implementation for the given runtime type.
+func newGatherer(rt types.RuntimeType) (gatherer, error) {
+	switch rt {
+	case types.RuntimeTypePodman:
+		return newPodmanGatherer(), nil
+	default:
+		return nil, fmt.Errorf("unsupported runtime for must-gather: %s", rt)
+	}
 }
